\name{cghE1}
\docType{data}
\alias{cghE1}
\title{A fictitious aCGH data set}
\description{
  A fictitious aCGH data set that contains UniGene identifiers to show
  how linking to IDClight \url{http://idclight.bioinfo.cnio.es} works.

}
\usage{cghE1}
\format{A data frame with 4323 rows and 9 columns, five of which
  correspond to the aCGH data for 5 samples.}
\source{Simulated data}
\references{
    Diaz-Uriarte, R. and Rueda, O.M. (2006). ADaCGH: an R package and
  web-based application for the analysis of aCGH data. Tech. report.
  \url{http://www.ligarto.org/rdiaz/Papers/adacgh.pdf},
  \url{http://adacgh.bioinfo.cnio.es}.
}
\keyword{datasets}