\name{cghMCRe}
\docType{data}
\alias{cghMCRe}
\title{A fictitious aCGH data set}
\description{
  A fictitious aCGH data set that can be used to show some Minimal
  common regions.

}
\usage{cghMCRe}
\format{A data frame with 4706 rows and 7 columns, three of which
  correspond to the aCGH data for 3 samples.}
\source{Simulated data}
\references{
    Diaz-Uriarte, R. and Rueda, O.M. (2006). ADaCGH: an R package and
  web-based application for the analysis of aCGH data. Tech. report.
  \url{http://www.ligarto.org/rdiaz/Papers/adacgh.pdf},
  \url{http://adacgh.bioinfo.cnio.es}.
}
\keyword{datasets}